requirements.txt
tensorflow==2.4.0
tensorflow-gpu==2.4.0
tqdm
pathlib
pandas
numpy
matplotlib
seaborn
pyyaml
opencv-python
Pillow
streamlit
jupyterlab
scikit-learn
scikit-image
sklearn
albumentations
ffmpeg
pytest
pytest-dependency
pre-commit
dvc[gs]
apache-airflow>=2.2.3
mlflow>=1.22.0
pyngrok